Transfer forma условно готово (без тестов)

+++ GPSS.Visualiztion/VisualGPSS_Schema.cs
@@ -95,7 +95,7 @@ namespace GPSS.Visualiztion
         }
 
         public void AddTransfer(string type, VisualBlock startBlock, VisualBlock block1,
-            VisualBlock block2, int digit, string label, string[] args, string comment)        
+            VisualBlock block2, double digit, string label, string[] args, string comment)        
         {
             Block.BlockType blockType = (Block.BlockType)Enum.Parse(typeof(Block.BlockType), type);
             Block block = new Block(label, args, comment, blockType);
